# swagger/swagger.yaml
basePath: /api/v0
definitions:
...
  consoleapi.NotificationListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/consoleapi.NotificationResponse'
        type: array
      limit:
        type: integer
      page:
        type: integer
      pageCount:
        type: integer
      totalCount:
        type: integer
    type: object
  consoleapi.NotificationResponse:
    properties:
      body:
        type: string
      createdAt:
        type: string
      data:
        additionalProperties: true
        type: object
      errorMessage:
        type: string
      hide:
        type: boolean
      id:
        type: string
      isRead:
        type: boolean
      retryCount:
        type: integer
      sentAt:
        type: string
      status:
        type: string
      title:
        type: string
      tokenId:
        type: string
      userId:
        type: string
    type: object
...
  consoleapi.SwaggerErrorResponse:
    properties:
      error:
        example: request validation failed
        type: string
    type: object
  consoleapi.UnreadCountResponse:
    properties:
      count:
        type: integer
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API documentation for StorX Monitor server
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: StorX Monitor API
  version: "1.0"
paths:
...
  /notifications:
    get:
      description: '**Full route:** `GET /api/v0/notifications`'
      parameters:
      - default: 50
        description: Page size (default 50)
        in: query
        name: limit
        type: integer
      - default: 1
        description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: 'Filter: unread'
        in: query
        name: filter
        type: string
      - description: 'Time window: 1d, 7d, 15d, 1m, 6m, 1y'
        in: query
        name: timeFilter
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/consoleapi.NotificationListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/consoleapi.SwaggerErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/consoleapi.SwaggerErrorResponse'
      security:
      - CookieAuth: []
      summary: List notifications
      tags:
      - notifications
  /notifications/{id}:
    get:
      description: '**Full route:** `GET /api/v0/notifications/{id}`'
      parameters:
      - description: Notification UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/consoleapi.NotificationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/consoleapi.SwaggerErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/consoleapi.SwaggerErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/consoleapi.SwaggerErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/consoleapi.SwaggerErrorResponse'
      security:
      - CookieAuth: []
      summary: Get notification detail
      tags:
      - notifications
  /notifications/{id}/dismiss:
    put:
      description: '**Full route:** `PUT /api/v0/notifications/{id}/dismiss`'
      parameters:
      - description: Notification UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/consoleapi.SwaggerErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/consoleapi.SwaggerErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/consoleapi.SwaggerErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/consoleapi.SwaggerErrorResponse'
      security:
      - CookieAuth: []
      summary: Dismiss notification
      tags:
      - notifications
  /notifications/count:
    get:
      description: '**Full route:** `GET /api/v0/notifications/count`'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/consoleapi.UnreadCountResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/consoleapi.SwaggerErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/consoleapi.SwaggerErrorResponse'
      security:
      - CookieAuth: []
      summary: Unread notification count
      tags:
      - notifications
  /notifications/read-all:
    put:
      description: '**Full route:** `PUT /api/v0/notifications/read-all`'
      produces:
      - application/json
      responses:
        "200":
          description: message
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/consoleapi.SwaggerErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/consoleapi.SwaggerErrorResponse'
      security:
      - CookieAuth: []
      summary: Mark all notifications read
      tags:
      - notifications
...
securityDefinitions:
  ApiKeyAuth:
    description: Bearer / API key (rare for console routes).
    in: header
    name: Authorization
    type: apiKey
  CSRFAuth:
    description: 'When `console.csrf-protection-enabled` is true: copy `csrfToken`
      from `GET /config` and Authorize here. Must match `csrf_token` cookie (set by
      GET /config or synced by Swagger UI). Required for `POST /auth/token`, `POST
      /auth/account/set-password`, and other CSRF-protected routes.'
    in: header
    name: X-CSRF-Token
    type: apiKey
  CookieAuth:
    description: Session cookie after login. **Swagger Authorize:** paste the token
      value only (from `POST /auth/token` response `token` field). Do not prefix with
      `Cookie:` or `_tokenKey=`. Copied curl uses `-b '_tokenKey=...'` (not a header).
      Login via Swagger auto-fills this.
    in: cookie
    name: _tokenKey
    type: apiKey
  SellerCSRFAuth:
    description: 'When `seller.csrf-protection-enabled` is true: copy `csrfToken`
      from `GET /seller/config` and Authorize here. Must match `csrf_token` cookie
      (set by GET /seller/config). Required for `POST /seller/auth/token`, MFA routes,
      password change, and other CSRF-protected seller routes.'
    in: header
    name: X-CSRF-Token
    type: apiKey
  SellerCookieAuth:
    description: Session cookie after reseller Google login. **Swagger Authorize:**
      paste the token value only (from `GET /seller/auth/google` response `token`
      field). Seller peer cookie name is `_seller_tokenKey` (not console `_tokenKey`).
    in: cookie
    name: _seller_tokenKey
    type: apiKey
swagger: "2.0"
...